Revert "Merge pull request #1400 from caveman2cosmos/realistic-culture-tweaks"
This reverts commit 5d43629b2c6355e4ac0e53fc5ddb0255f8a11878.

NOTE: Reverting master to pre-realistic culture merge, pending bugfixes

+++ Sources/CvUnit.h
@@ -2091,9 +2091,9 @@ protected:
 	void resolveAirCombat(CvUnit* pInterceptor, CvPlot* pPlot, CvAirMissionDefinition& kBattle);
 
 	// ------ BEGIN InfluenceDrivenWar -------------------------------
-	int doVictoryInfluence(CvUnit* pLoserUnit, bool bAttacking, bool bWithdrawal);
+	float doVictoryInfluence(CvUnit* pLoserUnit, bool bAttacking, bool bWithdrawal);
 	void influencePlots(CvPlot* pCentralPlot, const PlayerTypes eTargetPlayer, const int iLocationMultiplier);
-	int doPillageInfluence();
+	float doPillageInfluence();
 	// ------ END InfluenceDrivenWar ---------------------------------
 
 	const PromotionKeyedInfo*	findPromotionKeyedInfo(PromotionTypes ePromotion) const;
